Assignment02/requirements.txt
matplotlib==3.6.3
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
sklearn==1.4.1
gensim==4.3.2
pytrec-eval-terrier==0.5.6